Trait contec_protocol::AsyncReadWrite

pub trait AsyncReadWrite {
    type Error: AsErrorSource;

    fn poll_read(
        self: Pin<&mut Self>,
        cx: &mut Context<'_>,
        buf: &mut [u8]
    ) -> Poll<Result<usize, Self::Error>>;
    fn poll_write(
        self: Pin<&mut Self>,
        cx: &mut Context<'_>,
        buf: &[u8]
    ) -> Poll<Result<usize, Self::Error>>;
}

Trait that must be implemented by the serial port that enables async communication

Required Associated Types

type Error: AsErrorSource

Device error

Required Methods

fn poll_read(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
buf: &mut [u8]
) -> Poll<Result<usize, Self::Error>>

Attempt to read from the object into buf.

fn poll_write(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
buf: &[u8]
) -> Poll<Result<usize, Self::Error>>

Attempt to write bytes from buf into the object.